Moved classes to com.baeldung.validation.custommessage package
