+++ spring-boot-modules/spring-boot-custom-validation/src/main/java/com/baeldung/validation/custommessage/AppConfig.java
@@ -1,4 +1,4 @@
-package com.baeldung.validation;
+package com.baeldung.validation.custommessage;
 
 import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
