+++ spring-boot-modules/spring-boot-custom-validation/src/main/java/com/baeldung/validation/custommessage/UserDTO.java
@@ -1,4 +1,4 @@
-package com.baeldung.validation;
+package com.baeldung.validation.custommessage;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Min;
